back_utils/pothole-prediction/azure_back/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

--extra-index-url https://download.pytorch.org/whl/cpu

torch
torchvision
torchaudio
Pillow
azure-functions
azure-data-tables
azure-storage-blob
dotenv

